Компаратор убран в собственную модель
